+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.23] - 2025-12-18
+
 - Added new method `getMicrosoftAgreementValidationStatus` in CustomersClient to retrieve MCA status.
 - Added new method `getCredentials` in Licenses to handle credential requests.
 
@@ -653,7 +655,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Entity Price: use PriceBand now, which is in the Catalog namespace where it belongs
 - Entity Service: the root entity is deprecated, and the one in the Catalog namespace should be used
 
-[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.11.22...HEAD
+[Unreleased]: https://github.com/ArrowSphere/public-api-client/compare/0.11.23...HEAD
+[0.11.23]: https://github.com/ArrowSphere/public-api-client/compare/0.11.22...0.11.23
 [0.11.22]: https://github.com/ArrowSphere/public-api-client/compare/0.11.21...0.11.22
 [0.11.21]: https://github.com/ArrowSphere/public-api-client/compare/0.11.20...0.11.21
 [0.11.20]: https://github.com/ArrowSphere/public-api-client/compare/0.11.19...0.11.20